// h5/src/types/product.d.ts
export type PmsProduct = {
  /** 商品ID */
  id: number
  /** 品牌ID */
  brandId: number
  /** 品牌名称 */
  brandName: string
  /** 分类ID */
  productCategoryId: number
  /** 分类名称 */
  productCategoryName: string
  /** 商品名称 */
  name: string
  /** 图片 */
  pic: string
  /** 画册图片 */
  albumPics: string
  /** 上架状态：0->下架；1->上架 */
  publishStatus: number
  /** 新品状态：0->不是新品；1->新品 */
  newStatus: number
  /** 推荐状态：0->不推荐；1->推荐 */
  recommandStatus: number
  /** 审核状态：0->未审核；1->审核通过 */
  verifyStatus: number
  /** 排序 */
  sort: number
  /** 价格 */
  price: number
  /** 原价 */
  originalPrice: number
  /** 销量 */
  sale: number
  /** 库存 */
  stock: number
  /** 促销价格 */
  promotionPrice: number
  /** 促销类型：0->无促销；1->单品优惠；2->会员优惠；3->多买优惠；4->满减优惠；5->限时优惠 */
  promotionType: number
  /** 促销开始时间 */
  promotionStartTime: string
  /** 促销结束时间 */
  promotionEndTime: string
  /** 副标题 */
  subTitle: string
  /** 商品描述 */
  description: string
  /** 详情标题 */
  detailTitle: string
  /** 详情描述（移动端HTML） */
  detailMobileHtml: string
  /** 商品编码 */
  productSn: string
  /** 服务ID列表 */
  serviceIds: string
  /** 创建时间 */
  createTime: string
  /** 更新时间 */
  updateTime: string
  /** 秒杀价格 */
  flashPromotionPrice?: number
  /** 秒杀数量 */
  flashPromotionCount?: number
  /** 秒杀限购数量 */
  flashPromotionLimit?: number
}